:horse_racing: Better handling of procedure arguments
Now cpu_dispatch(...) takes varargs of Code instead of having multiple overloads to handle variable argument counts. The typechecking works as desired still since the arguments passed get assigned to a generated struct that has the types the procedure wants to take, so you can still do things like automatic casting, or complex expressions in an argument.

This allows us to greatly simplify the main codepath as well as support an arbitrary number of arguments. It also fixes possible issues with types not being referentiable by name on the module cause it might not have access to them.